Fix for multipart/signed
multipart/signed messages need to be created in two steps: first
part sets up the message as a multipart/signed message and creates
relevant parts.  The client must then be able to get the 'content'
(unsigned) part of the message including its headers to create the
hash signature, then in a second step, the client needs to set the
signature.